requirements.txt
azure-cognitiveservices-speech
python-dotenv
jieba
zhon
fastapi
uvicorn
python-multipart
streamlit
langchain
langchain-openai
langchain-core
openai
langchain-google-genai